wine-launcher: install script updated

+++ WineLauncher/install.sh
@@ -34,7 +34,13 @@ if [ "$default_install" = "yes" ]; then
  ln -s "$bin_dir/wine-launcher.sh" "$HOME/bin/wine-launcher.sh"
 fi
 
-if [ -d "$script_dir/extra" ]; then
- cp -rf "$script_dir/extra" "$bin_dir"
+rm -rf "$bin_dir/extra"
+
+if [ -d "$script_dir/extra" ] && [ -L "$script_dir/extra" ]; then
+ target=`readlink -e "$script_dir/extra"`
+ test ! -z "$target" && ln -s "$target" "$bin_dir/extra"
 fi
 
+if [ -d "$script_dir/extra" ] && [ ! -L "$script_dir/extra" ]; then
+ cp -rf "$script_dir/extra" "$bin_dir"
+fi
